+++ ctinspector/image.py
@@ -1,32 +1,11 @@
 from __future__ import print_function
-import tarfile
 import json
-from os.path import join, exists
-from os import makedirs
-from shutil import move
-from tempfile import gettempdir, NamedTemporaryFile, mkdtemp
+import ctinspector.layer
+from os.path import join
 from ctinspector.utils import print_list
 
-def extract(image):
-    image_cache_prefix = join(gettempdir(), "ctinspector", "image-cache")
-    image_cache_path = join(image_cache_prefix, image.id)
-    if not exists(image_cache_prefix):
-        makedirs(image_cache_prefix, 0o700)
-
-    # Image is not yet cached, save and extract it
-    if not exists(image_cache_path):
-        with NamedTemporaryFile(delete=False) as tmp_filename:
-            for chunk in image.save():
-                tmp_filename.write(chunk)
-        image_tar = tarfile.open(tmp_filename.name)
-        tmp_dir = mkdtemp()
-        image_tar.extractall(tmp_dir)
-        move(tmp_dir, image_cache_path)
-
-    return image_cache_path
-
-
-def show_info(image_content_path):
+def show_info(image_content_path, image_id):
+    image_id = image_id.split(":")[1]
     with open(join(image_content_path, 'manifest.json')) as manifest_file:
         manifest_data = manifest_file.read()
     image_manifest = json.loads(manifest_data)
@@ -44,4 +23,6 @@ def show_info(image_content_path):
                     created_by = created_by.strip("\n")
                     if created_by:
                         created_by_list.append(created_by)
-            print_list("History", created_by_list)
+            print_list("Build History", created_by_list)
+        for layer in manifest_item['Layers']:
+            ctinspector.layer.show_info(image_content_path, layer)