+++ ctinspector/report.py
@@ -1,8 +1,9 @@
 from __future__ import print_function
 import docker
+import ctinspector.tar
 import ctinspector.image
 import ctinspector.layer
-from ctinspector.colorhelper import info
+from ctinspector.colorhelper import info, print_info
 #from ctinspect.utils import print_list
 
 def image_info(image_name):
@@ -10,10 +11,11 @@ def image_info(image_name):
     try:
         image = docker_client.images.get(image_name)
     except docker.errors.ImageNotFound:
-        print("Image %s was not found, you can use:\ndocker pull %s" % (info(image_name), info(image_name)))
+        print("Image %s was not found, you can use:\n    docker pull %s" % (info(image_name), info(image_name)))
         exit(2)
-    image_path = ctinspector.image.extract(image)
-    ctinspector.image.show_info(image_path)
+    print_info("Id:", image.id)
+    image_path = ctinspector.tar.extract_image(image)
+    ctinspector.image.show_info(image_path, image.id)
 
 def container_info(container_id):
     print(container_id)