+++ ctinspector/utils.py
@@ -2,10 +2,27 @@ from __future__ import print_function
 import six
 from ctinspector.colorhelper import print_info, print_header, info, info_header, term_columns
 
-def print_list(label, items):
+def print_list_value(key, value):
+    if isinstance(value, (six.string_types, int)):
+        print('{0:>15}: {1}'.format(key, info(value)))
+    if isinstance(value, list):
+        print('{0:>15}:'.format(key))
+        for value_item in value:
+            if value_item in [None, ""]:
+                continue
+            print(' '*15, info(value_item))
+    if isinstance(value, dict):
+        print('{0:>15}:'.format(key))
+        for dict_key, dict_value in value.items():
+            if dict_value in [None, "", [], {}]:
+                continue
+            print(' '*16+'{0:>15}: {1}'.format(info_header(dict_key), info(dict_value)))
+
+def print_list(label, items, with_split=False):
     if not items:
         return
-    print_header("_" * term_columns())
+    if with_split:
+        print_header("_" * term_columns())
     print_header(label)
     if isinstance(items, (list, tuple)):
         for item in items:
@@ -13,14 +30,11 @@ def print_list(label, items):
     if isinstance(items, dict):
         for key, value in items.items():
             if value:
-                if isinstance(value, six.string_types):
-                    print('{0:>15}: {1}'.format(key, info(value)))
-                if isinstance(value, list):
-                    print('{0:>15}:'.format(key))
-                    for value_item in value:
-                        print(' '*15, info(value_item))
-                if isinstance(value, dict):
-                    print('{0:>15}:'.format(key))
-                    for dict_key, dict_value in value.items():
-                        print(' '*16+'{0:>15}: {1}'.format(info_header(dict_key), info(dict_value)))
-    print_header("_" * term_columns())
+                print_list_value(key, value)
+
+    if with_split:
+        print_header("_" * term_columns())
+
+def human_size(bytes, units=[' bytes', ' KB', ' MB', ' GB', ' TB', ' PB', ' EB']):  # pylint: disable=W0102
+    """ Returns a human readable string reprentation of bytes"""
+    return str(bytes) + units[0] if bytes < 1024 else human_size(bytes>>10, units[1:])